Complete merge of PipeInputDetails into PipeInputSpec
